pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "task_manager"
authors = ["Vladimir <vvtatarinoff@gmail.com>"]
packages = [{include = 'task_manager'}]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0"
python-dotenv = "^0.19.2"
django-bootstrap4 = "^21.2"
dj-database-url = "^0.5.0"
whitenoise = "^5.3.0"
rollbar = "^0.16.2"


[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
gunicorn = "^20.1.0"
pytest-django = "^4.5.2"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
